output-consumer/pyproject.toml
[project]
name = "output-consumer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch>=9.0.2",
    "kafka-python>=2.2.15",
]
